Bug fixes in utils.py Basins class. Last connective bathymetry distribution is now calculated.
Bathymetry.py class for reconstruction bathymetry using seafloor-age grids
now allows for different planetary thermal conditions (see Rosas and Korenaga
2021).

Bathymetry.py now allows for basin mergers described by a set of merger rules
and a description of how the rules should be applied. The rules are applied
in the Basins class (in utils.py). The set of rules for particular reconstructions
are also stored in a get function in the utils.py script. The set of rules
can be applied with Basins.applyMergeBasinMethods(...)

The merger rules written into utils.py Basins class are as follows:
1) mergeBasins:	Mergers basins given a a) parent and b) list of daughter basin.
2) mergeSmallBasins: Merges small basins with the closest large basin based on small basin threshold (say, percentage of total seafloor)
3) mergeSmallestConnection: Iteratively merges small and well connected basins parent basin until n basins are defined.

The connection strength to parent basin is calculated in Basins.calculateBasinConnectivityParameters(..) and is as follows:
(1/areaofBasinDaughter)*np.nansum( areaofConnection*connectiveBathymetryDistribution );

Jupyter notebooks have been updated to use the new methods described.